% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{as.data.frame.sunburstResult}
\alias{as.data.frame.sunburstResult}
\title{as.data.frame.sunburstResult}
\usage{
\method{as.data.frame}{sunburstResult}(x, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{x}{(sunburstResult) A sunburst treemap results object}

\item{...}{(none) Not used}

\item{stringsAsFactors}{(logical) Transform strings to factors, default FALSE}
}
\value{
Returns a data.frame
}
\description{
Coerces a sunburstResult object to data frame.
}
\seealso{
\code{\link{sunburstTreemap}} for generating the treemap that is
  the input for this function
}
