% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{as.data.frame.voronoiResult}
\alias{as.data.frame.voronoiResult}
\title{as.data.frame.voronoiResult}
\usage{
\method{as.data.frame}{voronoiResult}(x, ..., stringsAsFactors = FALSE)
}
\arguments{
\item{x}{(voronoiResult) A voronoi treemap results object}

\item{...}{(none) Not used}

\item{stringsAsFactors}{(logical) Transform strings to factors, default FALSE}
}
\value{
Returns a data.frame
}
\description{
Coerces a voronoiResult object to data frame, omitting polygon data.
}
\seealso{
\code{\link{voronoiTreemap}} for generating the treemap that is
  the input for this function
}
