% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cropped_voronoi}
\alias{cropped_voronoi}
\title{cropped_voronoi}
\usage{
cropped_voronoi(sites)
}
\arguments{
\item{sites}{The only input parameter for the function.
A set of XY coordinates that are used for tesselation}
}
\value{
A list of cell coordinates; one cell for each
  set of input coordinates.
}
\description{
Tesselates a plane using a set of XY coordinates
}
\details{
The function is only intended for
  internal use. However, one can also use it directly for 
  test purposes.
}
