% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawTreemap}
\alias{drawTreemap}
\title{drawTreemap}
\usage{
drawTreemap(treemap, add = FALSE, width = 0.9, height = 0.9,
  layout = c(1, 1), position = c(1, 1), cell_color = 1,
  color_palette = NULL, border_size = 6, border_color = grey(0.9),
  labels = lapply(treemap, function(x) x$level) \%>\% unlist \%>\% max,
  label_size = 1, label_color = grey(0.9), title = NULL)
}
\arguments{
\item{treemap}{(list) A list of grid polygons and text objects that is the
output from treemap generation.}

\item{add}{(logical) Defaults to \code{FALSE}, creating a new page when drawing
a treemap. When multiple treemaps should be plotted on the same page, this should be
set to TRUE, and position of treemaps specified by \code{layout} and \code{position} arguments.}

\item{width}{(numeric) The width (0 to 0.9) of the viewport that the treemap will occupy.}

\item{height}{(numeric) The height (0 to 0.9) of the viewport that the treemap will occupy.}

\item{layout}{(numeric) Vector of length 2 indicationg the number of rows and columns
that the plotting area is supposed to be subdivided in. Useful only together with
\code{position}, which indicates the position of the specific treemap. Use \code{add = TRUE}
to omit starting a new page every time you call \code{drawTreemap()}.}

\item{position}{(numeric) Vector of length 2 indicationg the position where the current
treemap should be drawn. Useful only together with \code{layout}, which indicates 
the number of rows and columns the plotting area is subdivided into. Use \code{add = TRUE}
to omit starting a new page every time you call \code{drawTreemap()}.}

\item{cell_color}{(numeric or character) A number representing the hierarchical level
that should be used for cell coloring (categorical), or \code{cell_size} for using relative
area to determine colors, or \code{custom_color} to use a color index supplied during
treemap generation, or \code{NULL} to omit drawing colors.}

\item{color_palette}{(character) A character vector of colors used to fill cells.
The default is to use \code{\link{rainbow_hcl}} from package \code{colorspace}}

\item{border_size}{(numeric) The initial line width of the highest level 
cells. Is reduced each level. Default is 6 pts.}

\item{border_color}{(character) Color for cell borders, default is a light grey.}

\item{labels}{(numeric) A number representing the hierarchical level that should be
used for drawing cell labels, or NULL to omit drawing labels. The default is the
lowest level (every cell has a label).}

\item{label_size}{(numeric) Expansion factor for labels. A simple multiplier
that shrinks or expands all labels, regardless of their relative size.}

\item{label_color}{(character) Color for cell labels.}

\item{title}{(character) An optional title, default to \code{NULL}.}
}
\value{
Creates a grid viewport and draws the treemap.
}
\description{
Draws the treemap from a list obtained by running [voronoiTreemap()] or
[sunburstTreemap()]. Many graphical parameters can be customized but some
settings that determine the appearance of treemaps are already made 
during treemap generation. Such parameters are primarily cell size and
and initial shape of the treemap.
}
\examples{
library(SysbioTreemaps)

# generate example data
df <- data.frame(
  A=rep(c("a", "b", "c"), each=15),
  B=sample(letters[4:13], 45, replace=TRUE),
  C=sample(1:100, 45)
)

# generate treemap
tm <- voronoiTreemap(
  data = df,
  levels = c("A", "B", "C"),
  cell_size = "C",
  cell_color = "A",
  maxIteration = 50,
  )

# plot treemap
drawTreemap(tm)

}
\seealso{
\code{\link{voronoiTreemap}} for generating the treemap (\code{list}) that is
  the input for the drawing function
}
