% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_centroids.R
\name{get_polygons}
\alias{get_polygons}
\title{get_polygons}
\usage{
get_polygons(x, ...)
}
\arguments{
\item{x}{(voronoiResult) A voronoi treemap results object}

\item{...}{(none) Not used}
}
\value{
List of polygons, as class POLYGON from sf
}
\description{
Extract the list of polygons from a voronoiResult object.
}
\seealso{
\code{\link{voronoiTreemap}} for generating the treemap that is
  the input for this function
}
