% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poly_sortpoints.R
\name{poly_sortpoints}
\alias{poly_sortpoints}
\title{poly_sortpoints}
\usage{
poly_sortpoints(df, x = "x", y = "y", clockwise = TRUE,
  vertex = NULL)
}
\arguments{
\item{df}{(data frame) A data.frame with X and Y coordinate columns}

\item{x}{(character) name of the X coordinate column}

\item{y}{(character) name of the Y coordinate column}

\item{clockwise}{(logical) sort points clockwise (default, TRUE) or 
counter-clockwise.}

\item{vertex}{(numeric) Vector of length 2 giving the centroid of the polygon.
Default is NULL, centroid is calculated from cordinates.}
}
\description{
Sort a set of points/coordinates in a clockwise or anti-clockwise
direction. For internal use. This function is now deprecated as 
its functionality has moved to the C++ core function, \code{cropped_voronoi()}.
}
\author{
original author Stuart K. Grange, modified by Michael Jahn
}
