% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_centroids.R
\name{poly_transform_shape}
\alias{poly_transform_shape}
\title{poly_transform_shape}
\usage{
poly_transform_shape(poly, xy_max = 2000)
}
\arguments{
\item{poly}{(list)  named list of 'x' and 'y' coordinates of the polygon}

\item{xy_max}{(numeric)  maximum desired expansion of the polygon (default: 2000)}
}
\value{
A list with slots 'x' and 'y' containing numeric coordinates
}
\description{
Mainly for internal use. The function transforms ('projects') arbitrary input
coordinates that are used as parent polygon or shape of the treemap.
By default it scales and centers polygon coordinates on a square
of 0 to 2000 units.
}
