% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{print.voronoiResult}
\alias{print.voronoiResult}
\title{print.voronoiResult}
\usage{
\method{print}{voronoiResult}(x, ...)
}
\arguments{
\item{x}{(voronoiResult) A voronoi treemap results object}

\item{...}{(none) Not used}
}
\value{
Returns a data.frame
}
\description{
Print method for voronoiResult
}
\seealso{
\code{\link{voronoiTreemap}} for generating the treemap that is
  the input for this function
}
