% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rounded_rect}
\alias{rounded_rect}
\title{Coordinates to draw a rounded rectangle as parent cell for treemaps}
\format{A data frame with 50 rows and 2 variables}
\usage{
data(Jahn_CellReports_2018)
}
\description{
\itemize{
  \item X - coordinate
  \item Y - coordinate
}
}
\keyword{datasets}
