% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{summary.sunburstResult}
\alias{summary.sunburstResult}
\title{summary.sunburstResult}
\usage{
\method{summary}{sunburstResult}(x, ...)
}
\arguments{
\item{x}{(sunburstResult) A sunburst treemap results object}

\item{...}{(none) Not used}
}
\value{
Returns a data.frame
}
\description{
Summary method for sunburstResult
}
\seealso{
\code{\link{sunburstTreemap}} for generating the treemap that is
  the input for this function
}
