% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_definitions.R
\name{summary.voronoiResult}
\alias{summary.voronoiResult}
\title{summary.voronoiResult}
\usage{
\method{summary}{voronoiResult}(x, ...)
}
\arguments{
\item{x}{(voronoiResult) A voronoi treemap results object}

\item{...}{(none) Not used}
}
\value{
Returns a data.frame
}
\description{
Summary method for voronoiResult.
}
\seealso{
\code{\link{voronoiTreemap}} for generating the treemap that is
  the input for this function
}
