% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawTreemap}
\alias{drawTreemap}
\title{drawTreemap}
\usage{
drawTreemap(treemap, width = 0.9, height = 0.9, main = NULL)
}
\arguments{
\item{data}{(data.frame) A data.frame with one column for each hierarchical level}
}
\value{
Creates a grid viewport and draws the treemap.
}
\description{
Draws the treemap from a list obtained by running [voronoiTreemap()]
}
\examples{

}
\seealso{
\code{\link{voronoiTreemap}} for generating the treemap (\code{list}) that is
  the input for the drawing function
}
