% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawTreemap}
\alias{drawTreemap}
\title{drawTreemap}
\usage{
drawTreemap(treemap, width = 0.9, height = 0.9, cell_color = 1,
  color_palette = NULL, border_size = 6, border_color = grey(0.9),
  labels = lapply(treemap, function(x) x$level) \%>\% unlist \%>\% max,
  label_color = grey(0.7), title = NULL)
}
\arguments{
\item{treemap}{(list) A list of grid polygons and text objects that is the
output from treemap generation.}

\item{width}{(numeric) The width (0 to 1) of the viewport that the treemap will occupy.}

\item{height}{(numeric) The height (0 to 1)of the viewport that the treemap will occupy.}

\item{cell_color}{(numeric or character) A number representing the hierarchical level
that should be used for cell coloring (categorical), or 'cell_size' for using relative
area to determine colors, or 'custom_color' to use a color index supplied during
treemap generation, or \code{NULL} to omit drawing colors.}

\item{color_palette}{(character) A character vector of colors used to fill cells.
The default is to use \code{\link{rainbow_hcl}} from package\code{colorspace}}

\item{border_size}{(numeric) The initial line width of the highest level 
cells. Is reduced each level. Default is 6 pts.}

\item{border_color}{(character) Color for cell borders, default is a light grey.}

\item{labels}{(numeric) A number representing the hierarchical level that should be
used for cell labels, or NULL to omit drawing labels. The default is the
lowest level (every cell get's a label).}

\item{label_color}{(character) Color for cell labels.}

\item{title}{(character) A title, optional.}
}
\value{
Creates a grid viewport and draws the treemap.
}
\description{
Draws the treemap from a list obtained by running [voronoiTreemap()] or
[sunburstTreemap()]. Many graphical parameters can be customized but some
settings that determine the appearance of treemaps are already made 
during treemap generation. Such parameters are primarily cell size and
and initial shape of the treemap.
}
\examples{
library(SysbioTreemaps)

# generate example data
df <- data.frame(
  A=rep(c("a", "b", "c"), each=15),
  B=sample(letters[4:13], 45, replace=TRUE),
  C=sample(1:100, 45)
)

# generate treemap
tm <- voronoiTreemap(
  data = df,
  levels = c("A", "B", "C"),
  cell_size = "C",
  cell_color = "A",
  maxIteration = 50,
  )

# plot treemap
drawTreemap(tm)

}
\seealso{
\code{\link{voronoiTreemap}} for generating the treemap (\code{list}) that is
  the input for the drawing function
}
