% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{drawTreemap}
\alias{drawTreemap}
\title{drawTreemap}
\usage{
drawTreemap(treemap, width = 0.9, height = 0.9, title = NULL)
}
\arguments{
\item{treemap}{(list) A list of grid polygons and text objects that is the
output from treemap generation.}

\item{width}{(numeric) The width (0 to 1) of the viewport that the treemap will occupy.}

\item{height}{(numeric) The height (0 to 1)of the viewport that the treemap will occupy.}

\item{title}{(character) A title, optional.}
}
\value{
Creates a grid viewport and draws the treemap.
}
\description{
Draws the treemap from a list obtained by running [voronoiTreemap()] or
[sunburstTreemap()]. Some graphical parameters can be tweaked but most
settings are made during treemap generation.
}
\examples{
library(SysbioTreemaps)

# generate example data
df <- data.frame(
  A=rep(c("a", "b", "c"), each=15),
  B=sample(letters[4:13], 45, replace=TRUE),
  C=sample(1:100, 45)
)

# generate treemap
tm <- voronoiTreemap(
  data = df,
  levels = c("A", "B", "C"),
  cell.size = "C",
  cell.color = "A",
  maxIteration = 50,
  )

# plot treemap
drawTreemap(tm)

}
\seealso{
\code{\link{voronoiTreemap}} for generating the treemap (\code{list}) that is
  the input for the drawing function
}
